Compute TES

Creates a structure for constructing the TES as described in "A Dynamical Model of Genetic Networks for Cell Differentiation Villani M, Barbieri A, Serra R (2011) A Dynamical Model of Genetic Networks for Cell Differentiation. PLOS ONE 6(3): e17703. https://doi.org/10.1371/journal.pone.0017703"

Usage

getTESs(ATM)

Arguments

ATM

ATM structure as returned from the getATM method.

Value

The output will be a named list that contains the list of computed TESs, the noise thresholds at which they emerged and lastly the ATM structure.

Examples

net <- BoolNet::generateRandomNKNetwork(10, 2)
attractors <- BoolNet::getAttractors(net)
ATM <- getATM(net, attractors)
getTESs(ATM)
#> $TES
#> $TES$level_0
#> $TES$level_0$TES_1
#>  [1] "a1"  "a2"  "a3"  "a4"  "a5"  "a6"  "a7"  "a8"  "a9"  "a10" "a11" "a12"
#> 
#> 
#> $TES$level_1
#> $TES$level_1$TES_2
#> [1] "a9"  "a10" "a11" "a12"
#> 
#> 
#> $TES$level_2
#> $TES$level_2$TES_3
#> [1] "a9"  "a10" "a11" "a12"
#> 
#> 
#> $TES$level_3
#> $TES$level_3$TES_4
#> [1] "a8"
#> 
#> $TES$level_3$TES_5
#> [1] "a9"  "a10" "a11" "a12"
#> 
#> 
#> $TES$level_4
#> $TES$level_4$TES_6
#> [1] "a1"
#> 
#> $TES$level_4$TES_7
#> [1] "a2"
#> 
#> $TES$level_4$TES_8
#> [1] "a3"
#> 
#> $TES$level_4$TES_9
#> [1] "a4"
#> 
#> $TES$level_4$TES_10
#> [1] "a5" "a6"
#> 
#> $TES$level_4$TES_11
#> [1] "a7"
#> 
#> $TES$level_4$TES_12
#> [1] "a8"
#> 
#> $TES$level_4$TES_13
#> [1] "a9"  "a11"
#> 
#> 
#> $TES$level_5
#> $TES$level_5$TES_14
#> [1] "a1"
#> 
#> $TES$level_5$TES_15
#> [1] "a2"
#> 
#> $TES$level_5$TES_16
#> [1] "a3"
#> 
#> $TES$level_5$TES_17
#> [1] "a4"
#> 
#> $TES$level_5$TES_18
#> [1] "a5"
#> 
#> $TES$level_5$TES_19
#> [1] "a6"
#> 
#> $TES$level_5$TES_20
#> [1] "a7"
#> 
#> $TES$level_5$TES_21
#> [1] "a8"
#> 
#> $TES$level_5$TES_22
#> [1] "a9"  "a11"
#> 
#> 
#> $TES$level_6
#> $TES$level_6$TES_23
#> [1] "a1"
#> 
#> $TES$level_6$TES_24
#> [1] "a2"
#> 
#> $TES$level_6$TES_25
#> [1] "a3"
#> 
#> $TES$level_6$TES_26
#> [1] "a4"
#> 
#> $TES$level_6$TES_27
#> [1] "a5"
#> 
#> $TES$level_6$TES_28
#> [1] "a6"
#> 
#> $TES$level_6$TES_29
#> [1] "a7"
#> 
#> $TES$level_6$TES_30
#> [1] "a8"
#> 
#> $TES$level_6$TES_31
#> [1] "a9"  "a11"
#> 
#> $TES$level_6$TES_32
#> [1] "a10"
#> 
#> $TES$level_6$TES_33
#> [1] "a12"
#> 
#> 
#> $TES$level_7
#> $TES$level_7$TES_34
#> [1] "a1"
#> 
#> $TES$level_7$TES_35
#> [1] "a2"
#> 
#> $TES$level_7$TES_36
#> [1] "a3"
#> 
#> $TES$level_7$TES_37
#> [1] "a4"
#> 
#> $TES$level_7$TES_38
#> [1] "a5"
#> 
#> $TES$level_7$TES_39
#> [1] "a6"
#> 
#> $TES$level_7$TES_40
#> [1] "a7"
#> 
#> $TES$level_7$TES_41
#> [1] "a8"
#> 
#> $TES$level_7$TES_42
#> [1] "a9"
#> 
#> $TES$level_7$TES_43
#> [1] "a10"
#> 
#> $TES$level_7$TES_44
#> [1] "a12"
#> 
#> 
#> $TES$level_8
#> $TES$level_8$TES_45
#> [1] "a1"
#> 
#> $TES$level_8$TES_46
#> [1] "a2"
#> 
#> $TES$level_8$TES_47
#> [1] "a3"
#> 
#> $TES$level_8$TES_48
#> [1] "a4"
#> 
#> $TES$level_8$TES_49
#> [1] "a5"
#> 
#> $TES$level_8$TES_50
#> [1] "a6"
#> 
#> $TES$level_8$TES_51
#> [1] "a7"
#> 
#> $TES$level_8$TES_52
#> [1] "a8"
#> 
#> $TES$level_8$TES_53
#> [1] "a9"
#> 
#> $TES$level_8$TES_54
#> [1] "a10"
#> 
#> $TES$level_8$TES_55
#> [1] "a11"
#> 
#> $TES$level_8$TES_56
#> [1] "a12"
#> 
#> 
#> 
#> $thresholds
#> [1] 0.00 0.03 0.05 0.10 0.15 0.20 0.30 0.35 0.40
#> 
#> $ATM
#> $ATM$ATM
#>       a1   a2   a3   a4   a5   a6   a7   a8   a9  a10  a11  a12
#> a1  0.30 0.15 0.10 0.00 0.10 0.00 0.10 0.00 0.15 0.00 0.10 0.00
#> a2  0.15 0.30 0.00 0.10 0.00 0.10 0.00 0.10 0.10 0.05 0.05 0.05
#> a3  0.10 0.00 0.30 0.15 0.10 0.00 0.10 0.00 0.15 0.00 0.10 0.00
#> a4  0.00 0.10 0.15 0.30 0.00 0.10 0.00 0.10 0.10 0.05 0.05 0.05
#> a5  0.10 0.00 0.10 0.00 0.40 0.20 0.00 0.00 0.15 0.00 0.05 0.00
#> a6  0.00 0.10 0.00 0.10 0.20 0.40 0.00 0.00 0.10 0.05 0.00 0.05
#> a7  0.10 0.00 0.10 0.00 0.00 0.00 0.40 0.10 0.15 0.00 0.15 0.00
#> a8  0.00 0.10 0.00 0.10 0.00 0.00 0.10 0.40 0.10 0.05 0.10 0.05
#> a9  0.03 0.00 0.03 0.00 0.03 0.00 0.03 0.00 0.40 0.15 0.35 0.00
#> a10 0.00 0.03 0.00 0.03 0.00 0.03 0.00 0.03 0.30 0.30 0.00 0.30
#> a11 0.03 0.00 0.03 0.00 0.03 0.00 0.03 0.00 0.40 0.00 0.35 0.15
#> a12 0.00 0.03 0.00 0.03 0.00 0.03 0.00 0.03 0.15 0.30 0.15 0.30
#> 
#> $ATM$lostFLips
#> [1] 0
#> 
#> $ATM$attractors
#> $ATM$attractors$decimal
#> $ATM$attractors$decimal$a1
#> $ATM$attractors$decimal$a1$involvedStates
#>      [,1] [,2]
#> [1,]   18  666
#> 
#> $ATM$attractors$decimal$a1$basinSize
#> [1] 24
#> 
#> 
#> $ATM$attractors$decimal$a2
#> $ATM$attractors$decimal$a2$involvedStates
#>      [,1] [,2]
#> [1,]   19  667
#> 
#> $ATM$attractors$decimal$a2$basinSize
#> [1] 24
#> 
#> 
#> $ATM$attractors$decimal$a3
#> $ATM$attractors$decimal$a3$involvedStates
#>      [,1] [,2]
#> [1,]   26  658
#> 
#> $ATM$attractors$decimal$a3$basinSize
#> [1] 24
#> 
#> 
#> $ATM$attractors$decimal$a4
#> $ATM$attractors$decimal$a4$involvedStates
#>      [,1] [,2]
#> [1,]   27  659
#> 
#> $ATM$attractors$decimal$a4$basinSize
#> [1] 24
#> 
#> 
#> $ATM$attractors$decimal$a5
#> $ATM$attractors$decimal$a5$involvedStates
#>      [,1] [,2]
#> [1,]  146  154
#> 
#> $ATM$attractors$decimal$a5$basinSize
#> [1] 32
#> 
#> 
#> $ATM$attractors$decimal$a6
#> $ATM$attractors$decimal$a6$involvedStates
#>      [,1] [,2]
#> [1,]  147  155
#> 
#> $ATM$attractors$decimal$a6$basinSize
#> [1] 32
#> 
#> 
#> $ATM$attractors$decimal$a7
#> $ATM$attractors$decimal$a7$involvedStates
#>      [,1] [,2]
#> [1,]  530  538
#> 
#> $ATM$attractors$decimal$a7$basinSize
#> [1] 16
#> 
#> 
#> $ATM$attractors$decimal$a8
#> $ATM$attractors$decimal$a8$involvedStates
#>      [,1] [,2]
#> [1,]  531  539
#> 
#> $ATM$attractors$decimal$a8$basinSize
#> [1] 16
#> 
#> 
#> $ATM$attractors$decimal$a9
#> $ATM$attractors$decimal$a9$involvedStates
#>      [,1] [,2] [,3] [,4]
#> [1,]    2  522  642  138
#> 
#> $ATM$attractors$decimal$a9$basinSize
#> [1] 464
#> 
#> 
#> $ATM$attractors$decimal$a10
#> $ATM$attractors$decimal$a10$involvedStates
#>      [,1] [,2] [,3] [,4]
#> [1,]    3  523  643  139
#> 
#> $ATM$attractors$decimal$a10$basinSize
#> [1] 112
#> 
#> 
#> $ATM$attractors$decimal$a11
#> $ATM$attractors$decimal$a11$involvedStates
#>      [,1] [,2] [,3] [,4]
#> [1,]   10  514  650  130
#> 
#> $ATM$attractors$decimal$a11$basinSize
#> [1] 144
#> 
#> 
#> $ATM$attractors$decimal$a12
#> $ATM$attractors$decimal$a12$involvedStates
#>      [,1] [,2] [,3] [,4]
#> [1,]   11  515  651  131
#> 
#> $ATM$attractors$decimal$a12$basinSize
#> [1] 112
#> 
#> 
#> 
#> $ATM$attractors$binary
#> $ATM$attractors$binary$a1
#>        [,1] [,2]
#> Gene1     0    0
#> Gene2     1    1
#> Gene3     0    0
#> Gene4     0    1
#> Gene5     1    1
#> Gene6     0    0
#> Gene7     0    0
#> Gene8     0    1
#> Gene9     0    0
#> Gene10    0    1
#> 
#> $ATM$attractors$binary$a2
#>        [,1] [,2]
#> Gene1     1    1
#> Gene2     1    1
#> Gene3     0    0
#> Gene4     0    1
#> Gene5     1    1
#> Gene6     0    0
#> Gene7     0    0
#> Gene8     0    1
#> Gene9     0    0
#> Gene10    0    1
#> 
#> $ATM$attractors$binary$a3
#>        [,1] [,2]
#> Gene1     0    0
#> Gene2     1    1
#> Gene3     0    0
#> Gene4     1    0
#> Gene5     1    1
#> Gene6     0    0
#> Gene7     0    0
#> Gene8     0    1
#> Gene9     0    0
#> Gene10    0    1
#> 
#> $ATM$attractors$binary$a4
#>        [,1] [,2]
#> Gene1     1    1
#> Gene2     1    1
#> Gene3     0    0
#> Gene4     1    0
#> Gene5     1    1
#> Gene6     0    0
#> Gene7     0    0
#> Gene8     0    1
#> Gene9     0    0
#> Gene10    0    1
#> 
#> $ATM$attractors$binary$a5
#>        [,1] [,2]
#> Gene1     0    0
#> Gene2     1    1
#> Gene3     0    0
#> Gene4     0    1
#> Gene5     1    1
#> Gene6     0    0
#> Gene7     0    0
#> Gene8     1    1
#> Gene9     0    0
#> Gene10    0    0
#> 
#> $ATM$attractors$binary$a6
#>        [,1] [,2]
#> Gene1     1    1
#> Gene2     1    1
#> Gene3     0    0
#> Gene4     0    1
#> Gene5     1    1
#> Gene6     0    0
#> Gene7     0    0
#> Gene8     1    1
#> Gene9     0    0
#> Gene10    0    0
#> 
#> $ATM$attractors$binary$a7
#>        [,1] [,2]
#> Gene1     0    0
#> Gene2     1    1
#> Gene3     0    0
#> Gene4     0    1
#> Gene5     1    1
#> Gene6     0    0
#> Gene7     0    0
#> Gene8     0    0
#> Gene9     0    0
#> Gene10    1    1
#> 
#> $ATM$attractors$binary$a8
#>        [,1] [,2]
#> Gene1     1    1
#> Gene2     1    1
#> Gene3     0    0
#> Gene4     0    1
#> Gene5     1    1
#> Gene6     0    0
#> Gene7     0    0
#> Gene8     0    0
#> Gene9     0    0
#> Gene10    1    1
#> 
#> $ATM$attractors$binary$a9
#>        [,1] [,2] [,3] [,4]
#> Gene1     0    0    0    0
#> Gene2     1    1    1    1
#> Gene3     0    0    0    0
#> Gene4     0    1    0    1
#> Gene5     0    0    0    0
#> Gene6     0    0    0    0
#> Gene7     0    0    0    0
#> Gene8     0    0    1    1
#> Gene9     0    0    0    0
#> Gene10    0    1    1    0
#> 
#> $ATM$attractors$binary$a10
#>        [,1] [,2] [,3] [,4]
#> Gene1     1    1    1    1
#> Gene2     1    1    1    1
#> Gene3     0    0    0    0
#> Gene4     0    1    0    1
#> Gene5     0    0    0    0
#> Gene6     0    0    0    0
#> Gene7     0    0    0    0
#> Gene8     0    0    1    1
#> Gene9     0    0    0    0
#> Gene10    0    1    1    0
#> 
#> $ATM$attractors$binary$a11
#>        [,1] [,2] [,3] [,4]
#> Gene1     0    0    0    0
#> Gene2     1    1    1    1
#> Gene3     0    0    0    0
#> Gene4     1    0    1    0
#> Gene5     0    0    0    0
#> Gene6     0    0    0    0
#> Gene7     0    0    0    0
#> Gene8     0    0    1    1
#> Gene9     0    0    0    0
#> Gene10    0    1    1    0
#> 
#> $ATM$attractors$binary$a12
#>        [,1] [,2] [,3] [,4]
#> Gene1     1    1    1    1
#> Gene2     1    1    1    1
#> Gene3     0    0    0    0
#> Gene4     1    0    1    0
#> Gene5     0    0    0    0
#> Gene6     0    0    0    0
#> Gene7     0    0    0    0
#> Gene8     0    0    1    1
#> Gene9     0    0    0    0
#> Gene10    0    1    1    0
#> 
#> 
#> 
#>